statfull_api/requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.6
django-debug-toolbar==2.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
drf-yasg==1.17.0
gunicorn==19.9.0
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
